backend/requirements.txt
# Orbit Backend - FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Validation & Settings
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0

# Authentication
pyjwt==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
google-auth==2.27.0
google-auth-oauthlib==1.2.0

# HTTP
httpx==0.26.0
aiohttp==3.9.1

# Redis (caching, rate limiting)
redis==5.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# Development
black==23.12.1
isort==5.13.2
mypy==1.8.0

# Gmail API
google-api-python-client==2.116.0

# ML/NLP
spacy==3.7.4
scikit-learn==1.4.0
rapidfuzz==3.6.1

# LLM
groq==0.4.2

# Encryption
cryptography==42.0.2

# Celery (background tasks)
celery[redis]==5.3.6
